develop in sqlite interface

+++ ProcessedHTTPReq.cpp
@@ -1,6 +1,6 @@
 #include <ProcessedHTTPReq.h>
 
-ProcessedHTTPReq::ProcessedHTTPReq(ClientInfo& client,string& host, monitor_time& time, long int no_pkt)
+ProcessedHTTPReq::ProcessedHTTPReq(ClientInfo& client,string& host, monitor_time& time, int no_pkt)
     : ProcessedPacket(client,host, time, no_pkt), requested()
 {
 }
